Added CrashDumpType to Schema
This adds the CrashDumpType field to the schema.
This setting enables you to specify the respective dump type for
collecting a crash. However, the way it's read by HCS, it's only
used to determine whether a crash should be collected or not.
If it's set to any valid value besides Disabled, crash dumps are
enabled.

Signed-off-by: Krithik Rao <krithikrao@microsoft.com>